requirements.txt
viam-sdk
pyserial